% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgplot3d_XQuartz.R
\name{bgplot3d_XQuartz}
\alias{bgplot3d_XQuartz}
\title{function for correcting new XQuartz issues}
\usage{
bgplot3d_XQuartz(expression)
}
\arguments{
\item{expression}{expression calls from DNE3d RFI3d and OPC3d
bgplot3d_XQuartz()}
}
\description{
crucial for plotting in Mac OS Yosemite and El Capitan
}

