% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered_patches.R
\name{clustered_patches}
\alias{clustered_patches}
\title{A clustering function}
\usage{
clustered_patches(Directional_Bin_Face_Pairs)
}
\arguments{
\item{Directional_Bin_Face_Pairs}{the bins of face directions
clustered_patches()}
}
\description{
This function gathers linked faces into patches
}

