% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_energy_per_face.R
\name{compute_energy_per_face}
\alias{compute_energy_per_face}
\title{Function will compute the DNE per face.}
\usage{
compute_energy_per_face(plyFile)
}
\arguments{
\item{plyFile}{a stanford PLY file 
compute_energy_per_face()}
}
\description{
This will generate each Dirichlet's normal energy for each triangular face on the surface.
}

