% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Directional_Bins.R
\name{Directional_Bins}
\alias{Directional_Bins}
\title{This bins the faces into directional categories}
\usage{
Directional_Bins(plyFile, rotation = 0)
}
\arguments{
\item{plyFile}{a stanford PLY file}

\item{rotation}{the amount to rotate the specimen by
Directional_Bins()}
}
\description{
bins into 8 directional categories on the basis of their orientations
}
