% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNE_Legend.R
\name{DNE_Legend}
\alias{DNE_Legend}
\title{Make legend for DNE3d plot}
\usage{
DNE_Legend(DNELabels, scaled = F, edgeMask = F, outlierMask = F,
  logColors = F, size = 1)
}
\arguments{
\item{DNELabels}{values for the labels}

\item{scaled}{logical indicating whether the values are scaled}

\item{edgeMask}{logical indicating whether of not edges are being masked
and that information to be included in the legend}

\item{outlierMask}{logical indicating whether outliers are masked}

\item{logColors}{logical indicating colors are on log scale}

\item{size}{legend scaling factor}
}
\description{
plotting subfunction
}
\details{
This is an internal function which builds a better DNE plot legend

The legend will reflect the elements used in the plot. This is an internal
function. Users will have little need or call to interact with it.
DNE_Legend()
}
