% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_vertices.R
\name{edge_vertices}
\alias{edge_vertices}
\title{Function for finding the edge vertices}
\usage{
edge_vertices(plyFile)
}
\arguments{
\item{plyFile}{a stanford PLY file


edge_vertices()}
}
\description{
Function will sort through all the vertices of the surface and find the ones which are on the edge. This will be needed for identifying which should be masked and not inlcuded in the calculation of the final DNE value.
}
