% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equal_Vertex_Normals.R
\name{Equal_Vertex_Normals}
\alias{Equal_Vertex_Normals}
\title{Important function for re-doing the vertex normals for the DNE calculation.}
\usage{
Equal_Vertex_Normals(plyFile)
}
\arguments{
\item{plyFile}{a stanford PLY file
Equal_Vertex_Normals()}
}
\description{
The geomorph import function does not generate the correct vertex normals.
}
