\name{ex_tooth1}
\docType{data}
\alias{ex_tooth1}
\title{4149_DU-LP-09_LM1}
\description{
	Lower M1 of a male mantled howler monkey, \emph{Aloutta palliata}. Catalogue Number DU-LP 09 
}
\usage{ex_tooth1}
\format{A list of five objects, as follows: \cr
	vb, a 4 x 5118 dataframe. \cr
	it, a 3 x 10000 dataframe. \cr
	primitivetype, a character string \cr
	material, a NULL field \cr
	normals, a 4 x 5118 dataframe
}
\source{MorphoSource}
\references{http://www.morphosource.com/index.php/Detail/SpecimenDetail/Show/specimen_id/22}
\keyword{datasets}