\name{ex_tooth2}
\docType{data}
\alias{ex_tooth2}
\title{4147_DU-LP-07_LM1}
\description{
	Lower M1 of a female manteled howler monkey, \emph{Aloutta palliata}. Catalogue Number DU-LP 07
}
\usage{ex_tooth1}
\format{A list of five objects, as follows: \cr
	vb, a 4 x 5135 dataframe. \cr
	it, a 3 x 9997 dataframe. \cr
	primitivetype, a character string \cr
	material, a NULL field \cr
	normals, a 4 x 5135 dataframe
}
\source{MorphoSource}
\references{http://www.morphosource.com/index.php/Detail/SpecimenDetail/Show/specimen_id/29}
\keyword{datasets}