% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/face_areas.R
\name{face_areas}
\alias{face_areas}
\title{Function to calculate face areas.}
\usage{
face_areas(plyFile)
}
\arguments{
\item{plyFile}{a stanford PLY file 
face_areas()}
}
\description{
This function calculates the area of each face on a ply file
}
