% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Face_Normals.R
\name{Face_Normals}
\alias{Face_Normals}
\title{Function to find Face Normals}
\usage{
Face_Normals(plyFile)
}
\arguments{
\item{plyFile}{a stanford PLY file
Face_Normals()}
}
\description{
This function re-computes the face normals in a way consistent with MorphoTester.
}

