% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_paired_directed_faces.R
\name{index_paired_directed_faces}
\alias{index_paired_directed_faces}
\title{Index of paired faces with directions}
\usage{
index_paired_directed_faces(plyFile)
}
\arguments{
\item{plyFile}{a stanford PLY file


index_paired_directed_faces()}
}
\description{
This does some heavy lifting to pull together faces which are paired together. This is needed for many later functions for compiling OPC
}

