% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molaR_Batch.R
\name{molaR_Batch}
\alias{molaR_Batch}
\title{Run a batch of molaR analyses}
\usage{
molaR_Batch(pathname = getwd(), DNE = TRUE, RFI = TRUE, OPCr = TRUE,
  OPC = FALSE, Details = FALSE, DNE_outliers = 0.1, RFI_alpha = 0.01,
  OPCr_steps = 8, OPCr_stepSize = 5.626, OPCr_minimum_faces = 3,
  OPCr_minimum_area = 0, OPC_rotation = 0, OPC_minimum_faces = 3,
  OPC_minimum_area = 0)
}
\arguments{
\item{pathname}{The path to the file containing all the PLY surfaces to be
analyzed. Defaults to the working directory}

\item{DNE}{logical indicating whether or not to perform DNE calculation Defaults
to true}

\item{RFI}{logical indicating whether or not to perform RFI calculation Defaults to
true}

\item{OPCr}{logical indicating whether or not to perform OPCr calculation
Defaults to true}

\item{OPC}{logical indicating whether or not to perform OPC calculation Defaults
to false}

\item{Details}{logical indicating whether or not to save the details of the RFI and
OPCr calculations}

\item{DNE_outliers}{the percentile at which outliers will be excluded is passed to
the DNE function, defaults to 0.1}

\item{RFI_alpha}{the size of the alpha passed to RFI
function, defaults to 0.01}

\item{OPCr_steps}{the number of steps the OPCr function should take, is passed to
the OPCr function. Defaults to 8}

\item{OPCr_stepSize}{the size of each rotation. Passed to the OPCr function.
Defaults to 5.626 degrees}

\item{OPCr_minimum_faces}{sets the lower boundary for number of faces a patch
must have for inclusion in total count. Defaults to 3 or more.}

\item{OPCr_minimum_area}{sets the lower boundary for percentage of the surface
area a patch must make up for inclusion in the total patch count. Cannot be used with
minimum_faces on. Defaults to zero}

\item{OPC_rotation}{amount of rotation to apply during OPC calculation. Defaults
to zero}

\item{OPC_minimum_faces}{minimum number of faces a patch must contain to be
counted in the OPC function. Defaults to 3.}

\item{OPC_minimum_area}{minimum percentage of the surface area a patch must
make up to be counted in the OPC function. Defaults to off}
}
\description{
A function which automats molaR analyses. User simply sets up the functions they
want run and can leave the computer to do the rest.
}
\details{
This function allows a user to set the analyses from molaR they want to run,
along with the specific parameters for each function and have a whole batch of PLY
files analyzed and saved to a csv file. Function will perform analyses on all PLY files
in the working directory or user can specify a file path.
}

