% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molaR_bgplot.R
\name{molaR_bgplot}
\alias{molaR_bgplot}
\title{Internal function for making molaR plot legends}
\usage{
molaR_bgplot(expression)
}
\arguments{
\item{expression}{it knows what to do...


molaR_bgplot()}
}
\description{
Function properly scales legends and prints them to the background of rgl
devices
}
