% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/patch_details.R
\name{patch_details}
\alias{patch_details}
\title{Function for gathering patch details for each Orientation patch}
\usage{
patch_details(clusterlist, plyFile)
}
\arguments{
\item{clusterlist}{a list of faces in the cluster
patch_details()}

\item{plyFile}{a stanford PLY file}
}
\description{
This function does some simple math to lets us know about the patches
}

