% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/patches_for_each_direction.R
\name{patches_for_each_direction}
\alias{patches_for_each_direction}
\title{Function for gathering the patches for each direction}
\usage{
patches_for_each_direction(indexed_pairs)
}
\arguments{
\item{indexed_pairs}{Pairs of touching faces}
}
\description{
This function will gather the patches in each of the 8 bins and ready it for
patches_for_each_direction()
}

