% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patches_per.R
\name{patches_per}
\alias{patches_per}
\title{A function for patches within each face}
\usage{
patches_per(patch_details, plyFile, minimum_faces = 3, minimum_area = 0)
}
\arguments{
\item{patch_details}{information on each patch}

\item{plyFile}{a stanford PLY file}

\item{minimum_faces}{minimum number of faces in each counted patch}

\item{minimum_area}{minimum area for a patch to be counted
patches_per()}
}
\description{
this gets some important information out of each patch
}

