% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/remove_boundary_faces.R
\name{remove_boundary_faces}
\alias{remove_boundary_faces}
\title{Remove boundary faces}
\usage{
remove_boundary_faces(Energy_Per_Face_Values, plyFile)
}
\arguments{
\item{Energy_Per_Face_Values}{information on E per face
remove_boundary_faces()}

\item{plyFile}{a stanford PLY file}
}
\description{
Important function for masking the edge faces
}

