% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_outliers.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Mask outliers on some faces}
\usage{
remove_outliers(Energy_values, X)
}
\arguments{
\item{Energy_values}{energy density values on faces}

\item{X}{percentile above which to remove


remove_outliers()}
}
\description{
This function will block out the top 0.1 percent of the faces
}
