% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repmat.R
\name{repmat}
\alias{repmat}
\title{Replicate a matrix}
\usage{
repmat(a, n, m = n)
}
\arguments{
\item{a}{vector or matrix to replicate}

\item{n}{number of replicates in n-direction}

\item{m}{number of replicates in m-direction


repmat}
}
\description{
This function replicates a matrix
}
