% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slope_Legend.R
\name{Slope_Legend}
\alias{Slope_Legend}
\title{Plotting subfunction for making slope plot legend}
\usage{
Slope_Legend(colors = colors, maskNegatives = T)
}
\arguments{
\item{colors}{provided from Slope3d}

\item{maskNegatives}{logical whether to mask}
}
\description{
Plotting subfunction for making slope plot legend
}
