% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molaR_package.R
\docType{data}
\name{Tooth}
\alias{Tooth}
\alias{Tooth.mesh}
\title{Tooth - scan of USNM_112176 lower M1 from Chlorocebus spp.}
\format{\code{Tooth}: triangular mesh representing a sine-cosine plane.}
\description{
A triangular mesh representing a lower M1 Chlorocebus spp. tooth - called by data(Tooth)
}
\keyword{datasets}
