% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertex_to_face_list.R
\name{vertex_to_face_list}
\alias{vertex_to_face_list}
\title{function for making a list of faces on each vertex}
\usage{
vertex_to_face_list(plyFile)
}
\arguments{
\item{plyFile}{a stanford PLY file
vertex_to_face_list()}
}
\description{
crucial function for getting a list of faces which will gather the faces per vertex.
}
