% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check2D.R
\name{Check2D}
\alias{Check2D}
\title{Plot 2D footprint and footprint triangle points to check for erros in 2D calculation}
\usage{
Check2D(RFI_Output, FootColor = "red", TriPointsColor = "black")
}
\arguments{
\item{RFI_Output}{An object that stores the output of the RFI
function}

\item{FootColor}{changes color of the 2D surface footprint}

\item{TriPointsColor}{color for the points of the footprint triangles}
}
\description{
This function will plot the points used for the 2D footprint area
calculation. This is meant to be a visual checking mechanism to
ensure that there are no 'extra' triangles within the footprint erroneously
adding to the total 2D area of the footprint. If a user finds extra points
within the boundaries of the footprint, they should assume that the alpha
value used for the RFI calculation was too small, and they are getting
a 2D footprint calculation which was too large.
}
\details{
This function will plot the points used for the 2D footprint area
calculation. This is meant to be a visual checking mechanism to
ensure that there are no 'extra' triangles within the footprint erroneously
adding to the total 2D area of the footprint. If a user finds extra points
within the boundaries of the footprint, they should assume that the alpha
value used for the RFI calculation was too small, and they are getting
a 2D footprint calculation which was too large.
}
