% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DNE.R
\name{DNE}
\alias{DNE}
\title{Calculate Dirichlet normal energy of a surface}
\usage{
DNE(plyFile)
}
\arguments{
\item{plyFile}{An object of class 'mesh3d' and 'shape3d' with calculated normals}
}
\description{
A function that calculates Dirichlet normal energy following the method of Bunn et
al. (2011) Comparing Dirichlet normal surface energy of tooth corwns, a new
technique of molar shape quantification for dietary inference, with previous methods
in isolation and in combination. Am J Phys Anthropol 145:247-261 doi: 10.1002
ajpa.21489
}
\details{
The function requires an object created by reading in a ply file utilizing
either the \code{read.ply} or the \code{read.AVIZO.ply} function, with calculated normals.

Dirichlet normal energy is calculated on meshes that represent specimen surfaces and
have already been downsampled to 10,000 faces and pre-smoothed in a 3D data
editing program.

The function does not include boundary faces in the calculation, and therefore the
analyzed surface cannot be closed (i.e., it must contain a hole). The function removes
the top 0.1 percent of calculated energy densities as outliers. Mesh orientation does not
affect this calculation.
}

