% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DNE3d.R
\name{DNE3d}
\alias{DNE3d}
\title{Plot results of a DNE analysis of a surface}
\usage{
DNE3d(DNE_File, setRange = c(0, 0), edgeMask = TRUE, outlierMask = TRUE,
  logColors = TRUE, showEdgePts = FALSE, fieldofview = 0, legend = TRUE)
}
\arguments{
\item{DNE_File}{An object that stores the output of the \code{DNE}
function}

\item{setRange}{User-defined range for plotting color scheme, see
Details}

\item{edgeMask}{Logical that colors edge faces black to indicate their
lack of contribution to the total Dirichlet normal energy}

\item{outlierMask}{Logical that colors outlier faces dark gray to
indicate their lack of contribution to the Dirichlet normal energy}

\item{logColors}{Logical that log transforms the color scheme}

\item{showEdgePts}{Logical that highlights the edge vertices in red
to indicate their lack of contribution to the total Dirichlet normal energy}

\item{fieldofview}{Passes an argument to \code{par3d} changing the field of
view in degrees of the resulting rgl}

\item{legend}{Logical indicating whether or not a legend
should be displayed}
}
\description{
A plotting function that colors the individual faces of a mesh by their relative 
contribution to the total Dirichlet normal energy calculated for the surface.
}
\details{
This function creates a heat map on the mesh surface
corresponding to the Dirichlet normal energy of each face calculated by
the \code{DNE} function. Hottest colors represent highest normal energy
values.

Dirichlet normal energies for the faces of a mesh surface tend to be
positively skewed, with a small proportion of the faces contributing
much of the total energy for the surface. When \code{logColors} is enabled the
function colorizes based on the log transformed Dirichlet normal
energies, allowing for finer resolution between faces near the mode of
the energy per face distribution. Disabling \code{logColors} will display the
untransformed Dirichlet normal energies.

The legend will update to reflect the other arguments chosen by the
user. Colors currently display in the legend in bins, however the colors
used in the displayed mesh surface are on a continuum. Ideally, the
legend should reflect a continuous stretch of color from the lowest
calculated Dirichlet normal energy to the highest. Future versions will
adjust the legend to this more intuitive display.

By default, the function sets the lowest Dirichlet normal energy
calculated among all faces to a cool color and the highest normal energy
calculated among all faces to red, and then colors the remaining faces
on a continuous color spectrum between these two end points using
either absolute or log transformed Dirichlet normal energy values
(depending on the status of \code{logColors}). Since the scale is relative to the
energies of the input surface, visual comparisons cannot directly be
made between  multiple plots of different surfaces. The code{setRange}
argument allows users to define the minimum and maximum of the
plotting color scheme and use it in multiple plots. This enables the
direct comparison of different surfaces to one another with red equal to
the user-defined maximum and a cool color equal to the user-defined
minimum. The user should choose reasonable bounds for the
maximum and minimum that are near the maximum and minimum
Dirichlet normal energies calculated for their surfaces. \code{setRange} will
not accept negative values.

\code{fieldofview} is set to a default of 0, which is an isometric projection.
Increasing it alters the degree of parallax in the perspective view, up to
a maximum of 179 degrees.
}

