% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molaR_package.R
\docType{data}
\name{Hills}
\alias{Hills}
\alias{Hills.mesh}
\title{Hills - sine-cosine plane}
\format{\code{Hills}: triangular mesh representing a sine-cosine plane.}
\description{
A triangular mesh representing a sine-cosine plane - called by data(Hills)
}
\keyword{datasets}
