% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPC_Legend.R
\name{OPC_Legend}
\alias{OPC_Legend}
\title{function for building a legend in OPC plots}
\usage{
OPC_Legend(binColors = c(1:8), binNumber = 8, maskDiscard = F,
  textCol = "black", size = 1, lineCol = "black")
}
\arguments{
\item{binColors}{number sequence for bins and their colors}

\item{binNumber}{numeric number of different bins}

\item{maskDiscard}{logical determining whether faces will be blacked out because
they are discarded}

\item{textCol}{color for the text in the circle legend}

\item{size}{scaling factor for the legend size}

\item{lineCol}{color for the lines in the legend
OPC_Legend()}
}
\description{
crucial graphics subfunction
}
