% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RFI3d.R
\name{RFI3d}
\alias{RFI3d}
\title{Plot 3D and 2D areas of a mesh used to calculate relief index}
\usage{
RFI3d(RFI_Output, displace = "Up", SurfaceColor = "gray",
  FootColor = "red", fieldofview = 0, Transparency = 1, legend = F)
}
\arguments{
\item{RFI_Output}{An object that stores the output of the \code{RFI}
function}

\item{displace}{Moves the surface footprint up, down, or not at
all}

\item{SurfaceColor}{changes the color of the 3D surface mesh}

\item{FootColor}{changes color of the 2D surface footprint}

\item{fieldofview}{Passes an argument to \code{par3d} changing the
field of view in degrees of the resulting rgl window}

\item{Transparency}{adjusts the transparency of the 3D mesh
surface}

\item{legend}{Logical indicating whether or not to include a
legend of the colors chosen to represent the 3D surface and
footprint}
}
\description{
A function that plots a three-dimensional model of the mesh
surface and includes a footprint of the two-dimensional area for
visual comparison.
}
\details{
This function can help to visualize the three-dimensional and two
dimensional areas that are used in calculating the relief index of a surface by
displaying both at the same time. The \code{RFI} function must be performed first.

\code{Transparency} can be adjusted in a range from fully opaque (1) to fully
transparent (0) in order to help visualize the footprint. The vertical placement of
the footprint along the Z axis can be altered with \code{displace} depending on how the
user wishes to view the surface, or on the original mesh orientation.

\code{fieldofview} is set to a default of 0, which is an isometric projection. Increasing it
alters the degree of parallax in the perspective view, up to a maximum of 179
degrees.
}

