% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Slope3d.R
\name{Slope3d}
\alias{Slope3d}
\title{Plot results of a Slope analysis of a surface}
\usage{
Slope3d(Slope_File, colors = c("red", "orangered", "orange", "yellow",
  "yellowgreen", "green", "cornflowerblue", "blue"), maskNegatives = T,
  leftOffset = 1, fieldofview = 0, legend = T)
}
\arguments{
\item{Slope_File}{An object that stores the output of the 
Slope function}

\item{colors}{A string of colors to build the color gradient}

\item{maskNegatives}{A logical indicating whether to mask 
the negative slopes or to reflect them into positive slopes}

\item{leftOffset}{numeric value between -1 and 1 setting 
the degree of offset for the plotted surface to the left. Larger 
values set further to right.}

\item{fieldofview}{Passes an argument to par3d changing
the field of view in degrees of the resulting rgl}

\item{legend}{Logical indicating whether or not a legend
shold be displayed}
}
\description{
plotting function
}
\details{
This function creates a heat map on the mesh surface
corresponding to the slope of each face calculated by
the Slope function. Colors are customizable
}

