% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tooth.R, R/molaR_package.R
\docType{data}
\name{Tooth}
\alias{Tooth}
\alias{Tooth.mesh}
\title{Tooth a surface mesh of a tooth.}
\format{
An object of class \code{"mesh3d"}

\code{Tooth}: triangular mesh representing a sine-cosine plane.
}
\usage{
data(Tooth)
}
\description{
Tooth scan of USNM_112176 lower M~1~ from \emph{Chlorocebus sp.}

A triangular mesh representing a lower M1 Chlorocebus spp. tooth - called by data(Tooth)
}
\examples{
data(Tooth)
DNE_Output <- DNE(Tooth)
DNE3d(DNE_Output)
}
\keyword{datasets}
